+++ src/performance/scala-2.11/markets/orderbooks/SortedOrderBookMicroBenchmark.scala
@@ -13,7 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package markets.engines.orderbooks
+package markets.orderbooks
 
 import markets.orders.AskOrder
 import markets.tradables.Tradable
@@ -23,8 +23,8 @@ import org.scalameter.{Bench, Gen}
 import scala.util.Random
 
 
-/** Performance tests for the `ConcurrentOrderBook` class. */
-object ConcurrentOrderBookMicroBenchmark extends Bench.OnlineRegressionReport {
+/** Performance tests for the `SortedOrderBook` class. */
+object SortedOrderBookMicroBenchmark extends Bench.OnlineRegressionReport {
 
   import markets.RandomOrderGenerator._
 
@@ -36,7 +36,7 @@ object ConcurrentOrderBookMicroBenchmark extends Bench.OnlineRegressionReport {
 
   /** Generates a collection of `ConcurrentOrderBook` instances of increasing size. */
   val orderBooks = for { size <- sizes } yield {
-    val orderBook = ConcurrentOrderBook[AskOrder](validTradable)
+    val orderBook = SortedOrderBook[AskOrder](validTradable)
     val orders = for (i <- 1 to size) yield randomAskOrder(prng, tradable = validTradable)
     orders.foreach( order => orderBook.add(order) )
     orderBook
@@ -49,7 +49,7 @@ object ConcurrentOrderBookMicroBenchmark extends Bench.OnlineRegressionReport {
     exec.jvmflags -> List("-Xmx2G")
     ) in {
 
-    /** Adding an `Order` to an `OrderBook` should be an `O(1)` operation. */
+    /** Adding an `Order` to a `SortedOrderBook` should be an `O(log n)` operation. */
     measure method "add" in {
       using(orderBooks) in {
         orderBook =>
@@ -58,7 +58,7 @@ object ConcurrentOrderBookMicroBenchmark extends Bench.OnlineRegressionReport {
       }
     }
 
-    /** Removing an `Order` from an `OrderBook` should be an `O(1)` operation. */
+    /** Removing an `Order` from a `SortedOrderBook` should be an `O(log n)` operation. */
     measure method "remove" in {
       using(orderBooks) in {
         orderBook =>
