+++ src/test/scala-2.11/markets/orderbooks/ConcurrentSortedOrderBookSpec.scala
@@ -13,7 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package markets.engines.orderbooks
+package markets.orderbooks
 
 import markets.RandomOrderGenerator._
 import markets.orders.{AskOrder, BidOrder}
